Refactor NeuralNetworkTextGenerator out to a different .js file
